+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## dev
+
+## [1.4.2] 29.06.2021
+
+- Add `redirectGuardName` [#24]
+- Implement `Replace`, `ReplaceName`, `replaceLastName` and `replaceLast`
+- Add `.QNotification.widgetBuilder` so the notification can be remove programmlcly.
+- Fix [#26]
+- Fix routing problem with child router.
+
 ## [1.4.1] 20.05.2021
 
 - Add `QR.settings.pagesType`